+++ index.js
@@ -3,4 +3,4 @@ require('dotenv').config();
 require('@babel/register');
 
 require('@babel/polyfill');
-require('./server/app');
+require('./server/server');
