+++ package.json
@@ -7,9 +7,9 @@
     "start": "nodemon index.js",
     "dev": "concurrently \"npm start\" ",
     "lint": "eslint .",
-    "test": "NODE_ENV=test jest --projects client server --detectOpenHandles",
+    "test": "NODE_ENV=test jest --projects client server",
     "test-client": "NODE_ENV=test jest --projects client",
-    "test-server": "NODE_ENV=test jest --projects server --forceExit",
+    "test-server": "NODE_ENV=test jest --projects server",
     "build": "webpack"
   },
   "author": "Various Authors",
