+++ server/__tests__/unit/routes/books.test.js
@@ -14,24 +14,44 @@ const request = require('supertest');
 // we also need to think about seperating our tests out into suites
 // so that we can run each on individually
 
+let server;
+  beforeAll(done => {
+    jest.useFakeTimers();
+    server = app.listen(5000, () => {
+      done();
+    });
+  });
+
+  afterAll(done => {
+    server.close(done);
+  });
+
 describe('GET /books ', () => {
-  test('/', async () => {
-    const response = await request(app).get('/api/books/');
-    expect(response.statusCode).toBe(200);
+  test('/', (done) => {
+    request(app)
+      .get('/api/books/')
+      .expect(200)
+      .end(done)
   });
-  test('/getBooksNoToken', async () => {
-    const response = await request(app).get('/api/books/getBooksNoToken');
-    expect(response.statusCode).toBe(200);
+  test('/getBooksNoToken', (done) => {
+    request(app)
+      .get('/api/books/getBooksNoToken')
+      .expect(200)
+      .end(done)
   });
-  test('/search', async () => {
-    const response = await request(app).get('/api/books/search');
-    expect(response.statusCode).toBe(200);
+  test('/search', (done) => {
+    request(app)
+      .get('/api/books/search')
+      .expect(200)
+      .end(done)
   });
 });
 
 describe('POST /books ', () => {
-  test('/requestBook should fail with no auth', async () => {
-    const response = await request(app).post('/api/books/requestBook');
-    expect(response.statusCode).toBe(401);
+  test('/requestBook should fail with no auth', (done) => {
+    request(app)
+      .post('/api/books/requestBook')
+      .expect(401)
+      .end(done)
   });
 });
